@@ -8,10 +8,9 @@ namespace Abp.AspNetCore.Configuration
 {
     public class ControllerAssemblySettingList : List<AbpControllerAssemblySetting>
     {
-        [CanBeNull]
-        public AbpControllerAssemblySetting GetSettingOrNull(Type controllerType)
+        public List<AbpControllerAssemblySetting> GetSettings(Type controllerType)
         {
-            return this.FirstOrDefault(controllerSetting => controllerSetting.Assembly == controllerType.GetAssembly());
+            return this.Where(controllerSetting => controllerSetting.Assembly == controllerType.GetAssembly()).ToList();
         }
     }
 }
