@@ -1,5 +1,6 @@
 using Abp.Dependency;
 using Microsoft.AspNetCore.Antiforgery;
+using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.Mvc.ViewFeatures.Internal;
 using Microsoft.Extensions.Logging;
@@ -9,15 +10,18 @@ namespace Abp.AspNetCore.Mvc.Antiforgery
 {
     public class AbpAutoValidateAntiforgeryTokenAuthorizationFilter : AutoValidateAntiforgeryTokenAuthorizationFilter, ITransientDependency
     {
-        private readonly AntiforgeryOptions _options;
+        private readonly AntiforgeryOptions _antiforgeryOptions;
+        private readonly CookieAuthenticationOptions _cookieAuthenticationOptions;
 
         public AbpAutoValidateAntiforgeryTokenAuthorizationFilter(
             IAntiforgery antiforgery, 
             ILoggerFactory loggerFactory,
-            IOptions<AntiforgeryOptions> options)
+            IOptions<AntiforgeryOptions> antiforgeryOptions,
+            IOptions<CookieAuthenticationOptions> cookieAuthenticationOptions)
             : base(antiforgery, loggerFactory)
         {
-            _options = options.Value;
+            _antiforgeryOptions = antiforgeryOptions.Value;
+            _cookieAuthenticationOptions = cookieAuthenticationOptions.Value;
         }
 
         protected override bool ShouldValidate(AuthorizationFilterContext context)
@@ -27,10 +31,17 @@ namespace Abp.AspNetCore.Mvc.Antiforgery
                 return false;
             }
 
+            //Always perform antiforgery validation when request contains authentication cookie
+            if (_cookieAuthenticationOptions.Cookie.Name != null &&
+                context.HttpContext.Request.Cookies.ContainsKey(_cookieAuthenticationOptions.Cookie.Name))
+            {
+                return true;
+            }
+
             //No need to validate if antiforgery cookie is not sent.
             //That means the request is sent from a non-browser client.
             //See https://github.com/aspnet/Antiforgery/issues/115
-            if (!context.HttpContext.Request.Cookies.ContainsKey(_options.CookieName))
+            if (!context.HttpContext.Request.Cookies.ContainsKey(_antiforgeryOptions.Cookie.Name))
             {
                 return false;
             }
