@@ -2,6 +2,7 @@ using System.Linq;
 using System.Reflection;
 using Abp.Application.Services;
 using Abp.AspNetCore.Configuration;
+using Abp.Collections.Extensions;
 using Abp.Dependency;
 using Abp.Reflection;
 using Microsoft.AspNetCore.Mvc.Controllers;
@@ -37,8 +38,8 @@ namespace Abp.AspNetCore.Mvc.Providers
                 return false;
             }
 
-            var configuration = _iocResolver.Resolve<AbpAspNetCoreConfiguration>().ControllerAssemblySettings.GetSettingOrNull(type);
-            return configuration != null && configuration.TypePredicate(type);
+            var settings = _iocResolver.Resolve<AbpAspNetCoreConfiguration>().ControllerAssemblySettings.GetSettings(type);
+            return settings.Any(setting => setting.TypePredicate(type));
         }
     }
 }
