@@ -15,17 +15,16 @@ namespace Abp.Json
     public class AbpMvcContractResolver : DefaultContractResolver
     {
         private readonly IIocResolver _iocResolver;
-        private bool _useMvcDateTimeFormat { get; set; }
+
+        private bool? _useMvcDateTimeFormat { get; set; }
+
         private string _datetimeFormat { get; set; } = null;
+
+        protected readonly object SyncObj = new object();
 
         public AbpMvcContractResolver(IIocResolver iocResolver)
         {
             _iocResolver = iocResolver;
-            using (var configuration = _iocResolver.ResolveAsDisposable<IAbpAspNetCoreConfiguration>())
-            {
-                _useMvcDateTimeFormat = configuration.Object.UseMvcDateTimeFormatForAppServices;
-            }
-
         }
 
         protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization memberSerialization)
@@ -44,27 +43,42 @@ namespace Abp.Json
                 return;
             }
 
-            if (ReflectionHelper.GetSingleAttributeOfMemberOrDeclaringTypeOrDefault<DisableDateTimeNormalizationAttribute>(member) == null)
+            if (ReflectionHelper.GetSingleAttributeOfMemberOrDeclaringTypeOrDefault<DisableDateTimeNormalizationAttribute>(member) != null)
             {
-                var converter = new AbpDateTimeConverter();
+                return;
+            }
 
-                // try to resolve MvcJsonOptions
-                if (_useMvcDateTimeFormat)
+            var converter = new AbpDateTimeConverter();
+
+            if (!_useMvcDateTimeFormat.HasValue)
+            {
+                lock (SyncObj)
                 {
-                    using (var mvcJsonOptions = _iocResolver.ResolveAsDisposable<IOptions<MvcJsonOptions>>())
+                    if (!_useMvcDateTimeFormat.HasValue)
                     {
-                        _datetimeFormat = mvcJsonOptions.Object.Value.SerializerSettings.DateFormatString;
+                        using (var configuration = _iocResolver.ResolveAsDisposable<IAbpAspNetCoreConfiguration>())
+                        {
+                            _useMvcDateTimeFormat = configuration.Object.UseMvcDateTimeFormatForAppServices;
+
+                            if (_useMvcDateTimeFormat.Value)
+                            {
+                                using (var mvcJsonOptions = _iocResolver.ResolveAsDisposable<IOptions<MvcJsonOptions>>())
+                                {
+                                    _datetimeFormat = mvcJsonOptions.Object.Value.SerializerSettings.DateFormatString;
+                                }
+                            }
+                        }
                     }
                 }
-
-                // apply DateTimeFormat only if not empty
-                if (!_datetimeFormat.IsNullOrWhiteSpace())
-                {
-                    converter.DateTimeFormat = _datetimeFormat;
-                }
-
-                property.Converter = converter;
             }
+            
+            // apply DateTimeFormat only if not empty
+            if (!_datetimeFormat.IsNullOrWhiteSpace())
+            {
+                converter.DateTimeFormat = _datetimeFormat;
+            }
+
+            property.Converter = converter;
         }
     }
 }
