@@ -1,13 +1,12 @@
-﻿using System;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-using Abp.Collections.Extensions;
+﻿using Abp.Collections.Extensions;
 using Abp.Domain.Entities;
 using Abp.Domain.Repositories;
 using NHibernate;
 using NHibernate.Linq;
-using NHibernate.Util;
+using System;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
 
 namespace Abp.NHibernate.Repositories
 {
