@@ -29,10 +29,10 @@ namespace Abp.IdentityFramework
                   {"User name '{0}' is invalid, can only contain letters or digits.", "Identity.InvalidUserName"},
                   {"A user with this login already exists.", "Identity.LoginAlreadyAssociated"},
                   {"Incorrect password.", "Identity.PasswordMismatch"},
-                  {"Passwords must have at least one digit ('0'-'9').", "Identity.PasswordRequiresDigit"},
-                  {"Passwords must have at least one lowercase ('a'-'z').", "Identity.PasswordRequiresLower"},
-                  {"Passwords must have at least one non alphanumeric character.", "Identity.PasswordRequiresNonAlphanumeric"},
-                  {"Passwords must have at least one uppercase ('A'-'Z').", "Identity.PasswordRequiresUpper"},
+                  {"Passwords must have at least one digit ('0'-'9').", "Identity.PasswordRequireDigit"},
+                  {"Passwords must have at least one lowercase ('a'-'z').", "Identity.PasswordRequireLower"},
+                  {"Passwords must have at least one non alphanumeric character.", "Identity.PasswordRequireNonAlphanumeric"},
+                  {"Passwords must have at least one uppercase ('A'-'Z').", "Identity.PasswordRequireUpper"},
                   {"Passwords must be at least {0} characters.", "Identity.PasswordTooShort"},
                   {"Role {0} does not exist.", "Identity.RoleNotFound"},
                   {"User already has a password set.", "Identity.UserAlreadyHasPassword"},
