@@ -10,8 +10,7 @@ namespace Abp.Localization
         /// <summary>
         /// Singleton instance.
         /// </summary>
-        public static NullLocalizationManager Instance { get { return SingletonInstance; } }
-        private static readonly NullLocalizationManager SingletonInstance = new NullLocalizationManager();
+        public static NullLocalizationManager Instance { get; } = new NullLocalizationManager();
 
         public LanguageInfo CurrentLanguage { get { return new LanguageInfo(CultureInfo.CurrentUICulture.Name, CultureInfo.CurrentUICulture.DisplayName); } }
 
@@ -21,7 +20,7 @@ namespace Abp.Localization
 
         private NullLocalizationManager()
         {
-            
+
         }
 
         public IReadOnlyList<LanguageInfo> GetAllLanguages()
