@@ -13,8 +13,7 @@ namespace Abp.Localization.Sources
         /// <summary>
         /// Singleton instance.
         /// </summary>
-        public static NullLocalizationSource Instance { get { return SingletonInstance; } }
-        private static readonly NullLocalizationSource SingletonInstance = new NullLocalizationSource();
+        public static NullLocalizationSource Instance { get; } = new NullLocalizationSource();
 
         public string Name { get { return null; } }
 
@@ -22,12 +21,12 @@ namespace Abp.Localization.Sources
 
         private NullLocalizationSource()
         {
-            
+
         }
 
         public void Initialize(ILocalizationConfiguration configuration, IIocResolver iocResolver)
         {
-            
+
         }
 
         public string GetString(string name)
