@@ -10,8 +10,7 @@ namespace Abp.Notifications
         /// <summary>
         /// Gets single instance of <see cref="NullRealTimeNotifier"/> class.
         /// </summary>
-        public static NullRealTimeNotifier Instance { get { return SingletonInstance; } }
-        private static readonly NullRealTimeNotifier SingletonInstance = new NullRealTimeNotifier();
+        public static NullRealTimeNotifier Instance { get; } = new NullRealTimeNotifier();
 
         public Task SendNotificationsAsync(UserNotification[] userNotifications)
         {
