@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Castle.Core.Logging;
 using Nito.AsyncEx;
@@ -85,6 +87,70 @@ namespace Abp.Runtime.Caching
             return item;
         }
 
+        public virtual object[] Get(string[] keys, Func<string, object> factory)
+        {
+            object[] items = null;
+
+            try
+            {
+                items = GetOrDefault(keys);
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex.ToString(), ex);
+            }
+
+            if (items == null)
+            {
+                items = new object[keys.Length];
+            }
+
+            if (items.Any(i => i == null))
+            {
+                lock (SyncObj)
+                {
+                    try
+                    {
+                        items = GetOrDefault(keys);
+                    }
+                    catch (Exception ex)
+                    {
+                        Logger.Error(ex.ToString(), ex);
+                    }
+
+                    var fetched = new List<KeyValuePair<string, object>>();
+                    for (var i = 0; i < items.Length; i++)
+                    {
+                        string key = keys[i];
+                        object value = items[i];
+                        if (value == null)
+                        {
+                            value = factory(key);
+                        }
+
+                        if (value != null)
+                        {
+                            fetched.Add(new KeyValuePair<string, object>(key, value));
+                        }
+                    }
+
+                    if (fetched.Any())
+                    {
+                        try
+                        {
+                            Set(fetched.ToArray());
+                        }
+                        catch (Exception ex)
+                        {
+                            Logger.Error(ex.ToString(), ex);
+                        }
+                    }
+                }
+            }
+
+            return items;
+        }
+
         public virtual async Task<object> GetAsync(string key, Func<string, Task<object>> factory)
         {
             object item = null;
@@ -135,27 +201,127 @@ namespace Abp.Runtime.Caching
             return item;
         }
 
+        public virtual async Task<object[]> GetAsync(string[] keys, Func<string, Task<object>> factory)
+        {
+            object[] items = null;
+
+            try
+            {
+                items = await GetOrDefaultAsync(keys);
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex.ToString(), ex);
+            }
+
+            if (items == null)
+            {
+                items = new object[keys.Length];
+            }
+
+            if (items.Any(i => i == null))
+            {
+                using (await _asyncLock.LockAsync())
+                {
+                    try
+                    {
+                        items = await GetOrDefaultAsync(keys);
+                    }
+                    catch (Exception ex)
+                    {
+                        Logger.Error(ex.ToString(), ex);
+                    }
+
+                    var fetched = new List<KeyValuePair<string, object>>();
+                    for (var i = 0; i < items.Length; i++)
+                    {
+                        string key = keys[i];
+                        object value = items[i];
+                        if (value == null)
+                        {
+                            value = factory(key);
+                        }
+
+                        if (value != null)
+                        {
+                            fetched.Add(new KeyValuePair<string, object>(key, value));
+                        }
+                    }
+
+                    if (fetched.Any())
+                    {
+                        try
+                        {
+                            await SetAsync(fetched.ToArray());
+                        }
+                        catch (Exception ex)
+                        {
+                            Logger.Error(ex.ToString(), ex);
+                        }
+                    }
+                }
+            }
+
+            return items;
+        }
+
         public abstract object GetOrDefault(string key);
+
+        public virtual object[] GetOrDefault(string[] keys)
+        {
+            return keys.Select(GetOrDefault).ToArray();
+        }
 
         public virtual Task<object> GetOrDefaultAsync(string key)
         {
             return Task.FromResult(GetOrDefault(key));
         }
 
+        public virtual Task<object[]> GetOrDefaultAsync(string[] keys)
+        {
+            return Task.FromResult(GetOrDefault(keys));
+        }
+
         public abstract void Set(string key, object value, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null);
+
+        public virtual void Set(KeyValuePair<string, object>[] pairs, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null)
+        {
+            foreach (var pair in pairs)
+            {
+                Set(pair.Key, pair.Value, slidingExpireTime, absoluteExpireTime);
+            }
+        }
 
         public virtual Task SetAsync(string key, object value, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null)
         {
-            Set(key, value, slidingExpireTime);
+            Set(key, value, slidingExpireTime, absoluteExpireTime);
             return Task.FromResult(0);
         }
 
+        public virtual Task SetAsync(KeyValuePair<string, object>[] pairs, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null)
+        {
+            return Task.WhenAll(pairs.Select(p => SetAsync(p.Key, p.Value, slidingExpireTime, absoluteExpireTime)));
+        }
+
         public abstract void Remove(string key);
+
+        public virtual void Remove(string[] keys)
+        {
+            foreach (var key in keys)
+            {
+                Remove(key);
+            }
+        }
 
         public virtual Task RemoveAsync(string key)
         {
             Remove(key);
             return Task.FromResult(0);
+        }
+
+        public virtual Task RemoveAsync(string[] keys)
+        {
+            return Task.WhenAll(keys.Select(RemoveAsync));
         }
 
         public abstract void Clear();
