@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Abp.Runtime.Caching
@@ -37,6 +38,16 @@ namespace Abp.Runtime.Caching
         object Get(string key, Func<string, object> factory);
 
         /// <summary>
+        /// Gets items from the cache.
+        /// This method hides cache provider failures (and logs them),
+        /// uses the factory method to get the object if cache provider fails.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <param name="factory">Factory method to create cache item if not exists</param>
+        /// <returns>Cached item</returns>
+        object[] Get(string[] keys, Func<string, object> factory);
+
+        /// <summary>
         /// Gets an item from the cache.
         /// This method hides cache provider failures (and logs them),
         /// uses the factory method to get the object if cache provider fails.
@@ -47,6 +58,16 @@ namespace Abp.Runtime.Caching
         Task<object> GetAsync(string key, Func<string, Task<object>> factory);
 
         /// <summary>
+        /// Gets items from the cache.
+        /// This method hides cache provider failures (and logs them),
+        /// uses the factory method to get the object if cache provider fails.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <param name="factory">Factory method to create cache item if not exists</param>
+        /// <returns>Cached items</returns>
+        Task<object[]> GetAsync(string[] keys, Func<string, Task<object>> factory);
+
+        /// <summary>
         /// Gets an item from the cache or null if not found.
         /// </summary>
         /// <param name="key">Key</param>
@@ -54,11 +75,25 @@ namespace Abp.Runtime.Caching
         object GetOrDefault(string key);
 
         /// <summary>
+        /// Gets items from the cache. For every key that is not found, a null value is returned.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <returns>Cached items</returns>
+        object[] GetOrDefault(string[] keys);
+
+        /// <summary>
         /// Gets an item from the cache or null if not found.
         /// </summary>
         /// <param name="key">Key</param>
         /// <returns>Cached item or null if not found</returns>
         Task<object> GetOrDefaultAsync(string key);
+
+        /// <summary>
+        /// Gets items from the cache. For every key that is not found, a null value is returned.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <returns>Cached items</returns>
+        Task<object[]> GetOrDefaultAsync(string[] keys);
 
         /// <summary>
         /// Saves/Overrides an item in the cache by a key.
@@ -74,6 +109,18 @@ namespace Abp.Runtime.Caching
         void Set(string key, object value, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null);
 
         /// <summary>
+        /// Saves/Overrides items in the cache by the pairs.
+        /// Use one of the expire times at most (<paramref name="slidingExpireTime"/> or <paramref name="absoluteExpireTime"/>).
+        /// If none of them is specified, then
+        /// <see cref="DefaultAbsoluteExpireTime"/> will be used if it's not null. Othewise, <see cref="DefaultSlidingExpireTime"/>
+        /// will be used.
+        /// </summary>
+        /// <param name="pairs">Pairs</param>
+        /// <param name="slidingExpireTime">Sliding expire time</param>
+        /// <param name="absoluteExpireTime">Absolute expire time</param>
+        void Set(KeyValuePair<string, object>[] pairs, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null);
+
+        /// <summary>
         /// Saves/Overrides an item in the cache by a key.
         /// Use one of the expire times at most (<paramref name="slidingExpireTime"/> or <paramref name="absoluteExpireTime"/>).
         /// If none of them is specified, then
@@ -87,16 +134,40 @@ namespace Abp.Runtime.Caching
         Task SetAsync(string key, object value, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null);
 
         /// <summary>
-        /// Removes a cache item by it's key.
+        /// Saves/Overrides items in the cache by the pairs.
+        /// Use one of the expire times at most (<paramref name="slidingExpireTime"/> or <paramref name="absoluteExpireTime"/>).
+        /// If none of them is specified, then
+        /// <see cref="DefaultAbsoluteExpireTime"/> will be used if it's not null. Othewise, <see cref="DefaultSlidingExpireTime"/>
+        /// will be used.
+        /// </summary>
+        /// <param name="pairs">Pairs</param>
+        /// <param name="slidingExpireTime">Sliding expire time</param>
+        /// <param name="absoluteExpireTime">Absolute expire time</param>
+        Task SetAsync(KeyValuePair<string, object>[] pairs, TimeSpan? slidingExpireTime = null, TimeSpan? absoluteExpireTime = null);
+
+        /// <summary>
+        /// Removes a cache item by it's key (does nothing if given key does not exists in the cache).
         /// </summary>
         /// <param name="key">Key</param>
         void Remove(string key);
+
+        /// <summary>
+        /// Removes cache items by their keys.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        void Remove(string[] keys);
 
         /// <summary>
         /// Removes a cache item by it's key (does nothing if given key does not exists in the cache).
         /// </summary>
         /// <param name="key">Key</param>
         Task RemoveAsync(string key);
+
+        /// <summary>
+        /// Removes cache items by their keys.
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        Task RemoveAsync(string[] keys);
 
         /// <summary>
         /// Clears all items in this cache.
