@@ -56,7 +56,7 @@ namespace Abp.AutoMapper.Tests
         }
 
         [Fact]
-        public void Should_Map_Two_Way_When_AutoMAp_Attribute_Is_Used()
+        public void Should_Map_Two_Way_When_AutoMap_Attribute_Is_Used()
         {
             MyClass3 obj2 = new MyClass3
             {
