@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 using Abp.Dapper.Repositories;
 using Abp.Dapper.Tests.Entities;
@@ -20,6 +21,8 @@ namespace Abp.Dapper.Tests
         private readonly IUnitOfWorkManager _unitOfWorkManager;
         private readonly IRepository<ProductDetail> _productDetailRepository;
         private readonly IDapperRepository<ProductDetail> _productDetailDapperRepository;
+        private readonly IRepository<Person> _personRepository;
+        private readonly IDapperRepository<Person> _personDapperRepository;
 
         public DapperRepository_Tests()
         {
@@ -28,10 +31,12 @@ namespace Abp.Dapper.Tests
             _unitOfWorkManager = Resolve<IUnitOfWorkManager>();
             _productDetailRepository = Resolve<IRepository<ProductDetail>>();
             _productDetailDapperRepository = Resolve<IDapperRepository<ProductDetail>>();
+            _personRepository = Resolve<IRepository<Person>>();
+            _personDapperRepository = Resolve<IDapperRepository<Person>>();
         }
 
         [Fact]
-        public void Dapper_Repository_Tests()
+        public async Task Dapper_Repository_Tests()
         {
             using (IUnitOfWorkCompleteHandle uow = _unitOfWorkManager.Begin())
             {
@@ -120,6 +125,9 @@ namespace Abp.Dapper.Tests
                 Product productWithTenantId3FromDapper = _productDapperRepository.FirstOrDefault(x => x.Name == "ProductWithTenant3");
                 productWithTenantId3FromDapper.ShouldBeNull();
 
+                Product p = await _productDapperRepository.FirstOrDefaultAsync(x => x.Status == Status.Active);
+                p.ShouldNotBeNull();
+
                 using (_unitOfWorkManager.Current.SetTenantId(3))
                 {
                     Product productWithTenantId3FromDapperInsideTenantScope = _productDapperRepository.FirstOrDefault(x => x.Name == "ProductWithTenant3");
@@ -137,14 +145,30 @@ namespace Abp.Dapper.Tests
                     productWithTenant40.CreatorUserId.ShouldBe(AbpSession.UserId);
                 }
 
-
                 //Second DbContext tests
-                int productDetailId =_productDetailRepository.InsertAndGetId(new ProductDetail("Woman"));
+                int productDetailId = _productDetailRepository.InsertAndGetId(new ProductDetail("Woman"));
                 _productDetailDapperRepository.Get(productDetailId).ShouldNotBeNull();
-
 
                 uow.Complete();
             }
+        }
+        //About issue-#3990
+        [Fact]
+        public void Should_Insert_Only_Have_IMustHaveTenant()
+        {
+            using (IUnitOfWorkCompleteHandle uow = _unitOfWorkManager.Begin())
+            {
+                using (_unitOfWorkManager.Current.SetTenantId(AbpSession.TenantId))
+                {
+                    int personWithTenantId40 = _personDapperRepository.InsertAndGetId(new Person("PersonWithTenantId40"));
+
+                    Person personWithTenant40 = _personRepository.Get(personWithTenantId40);
+
+                    personWithTenant40.TenantId.ShouldBe(AbpSession.TenantId.Value);
+                }
+
+            }
+
         }
     }
 }
