@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Abp.Authorization;
-using Abp.Authorization.Users;
 using Abp.Configuration;
+using Abp.Dependency;
 using Abp.Threading;
 using Abp.Zero.Configuration;
 using Abp.Zero.SampleApp.Authorization;
@@ -82,8 +82,46 @@ namespace Abp.Zero.SampleApp.Tests.Users
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public async Task Test_IsLockoutEnabled_On_User_Creation(bool isLockoutEnabled)
+        public async Task Test_IsLockoutEnabled_On_User_Creation_Should_Use_Setting_By_Default(bool isLockoutEnabledByDefault)
         {
+            // Arrange
+
+            ChangeLockoutEnabledSetting(isLockoutEnabledByDefault);
+
+            // Act
+
+            var user = new User
+            {
+                TenantId = AbpSession.TenantId,
+                UserName = "user1",
+                Name = "John",
+                Surname = "Doe",
+                EmailAddress = "user1@aspnetboilerplate.com",
+                IsEmailConfirmed = true,
+                Password = "AM4OLBpptxBYmM79lGOX9egzZk3vIQU3d/gFCJzaBjAPXzYIK3tQ2N7X4fcrHtElTw==", //123qwe
+                // IsLockoutEnabled = isLockoutEnabled
+            };
+
+            await WithUnitOfWorkAsync(async () => await _userManager.CreateAsync(user));
+
+            // Assert
+
+            (await _userManager.GetLockoutEnabledAsync(user.Id)).ShouldBe(isLockoutEnabledByDefault);
+        }
+
+        [Theory]
+        [InlineData(true, true, true)]
+        [InlineData(true, false, true)]
+        [InlineData(false, true, true)]
+        [InlineData(false, false, false)]
+        public async Task Test_IsLockoutEnabled_On_User_Creation_Should_Use_Stricter_Value_If_Set(bool isLockoutEnabledByDefault, bool isLockoutEnabled, bool isLockoutEnabledExpected)
+        {
+            // Arrange
+
+            ChangeLockoutEnabledSetting(isLockoutEnabledByDefault);
+
+            // Act
+
             var user = new User
             {
                 TenantId = AbpSession.TenantId,
@@ -98,7 +136,29 @@ namespace Abp.Zero.SampleApp.Tests.Users
 
             await WithUnitOfWorkAsync(async () => await _userManager.CreateAsync(user));
 
-            (await _userManager.GetLockoutEnabledAsync(user.Id)).ShouldBe(isLockoutEnabled);
+            // Assert
+
+            (await _userManager.GetLockoutEnabledAsync(user.Id)).ShouldBe(isLockoutEnabledExpected);
         }
+
+        #region Helpers
+
+        private void ChangeLockoutEnabledSetting(bool isLockoutEnabledByDefault)
+        {
+
+            LocalIocManager.Using<ISettingManager>(settingManager =>
+            {
+                if (AbpSession.TenantId is int tenantId)
+                {
+                    settingManager.ChangeSettingForTenant(tenantId, AbpZeroSettingNames.UserManagement.UserLockOut.IsEnabled, isLockoutEnabledByDefault.ToString());
+                }
+                else
+                {
+                    settingManager.ChangeSettingForApplication(AbpZeroSettingNames.UserManagement.UserLockOut.IsEnabled, isLockoutEnabledByDefault.ToString());
+                }
+            });
+        }
+
+        #endregion
     }
 }
