@@ -6,10 +6,19 @@ namespace Abp.ZeroCore.SampleApp.Application.Shop
     {
         public decimal Price { get; set; }
 
-        public int Stock { get; set; }
+        public string Name { get; set; }
+
+        public string Language { get; set; }
+    }
+
+    public class OrderListDto : EntityDto
+    {
+        public decimal Price { get; set; }
 
         public string Name { get; set; }
 
         public string Language { get; set; }
+
+        public int ProductCount { get; set; }
     }
 }
